♻️ Add file for CustomizableParser class
